+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased] - ReleaseDate
 
+## [0.19.0] - 2024-01-15
+
 ### Change
 
 - Add `Debug` support for `Xot` struct.
@@ -452,8 +454,9 @@
 Initial public release.
 
 <!-- next-url -->
+[Unreleased]: https://github.com/faassen/xot/compare/v0.19.0...HEAD
 
-[Unreleased]: https://github.com/faassen/xot/compare/v0.18.0...HEAD
+[0.19.0]: https://github.com/faassen/xot/compare/v0.18.0...v0.19.0
 [0.18.0]: https://github.com/faassen/xot/compare/v0.17.0...v0.18.0
 [0.17.0]: https://github.com/faassen/xot/compare/v0.16.0...v0.17.0
 [0.16.0]: https://github.com/faassen/xot/compare/v0.15.0...v0.16.0